+++ lib/intl_messages.dart
@@ -6,3 +6,4 @@ export 'src/intl_messages_basic_dictionary.dart';
 export 'src/intl_messages_date.dart';
 export 'src/intl_resource.dart';
 export 'src/locales.dart';
+export 'src/translator.dart';